feat!: Deprecate older modules
BREAKING CHANGE: This removes several older modules that may still be in use. Consult the deprecation warnings and documentation to find the new modules to use.
