+++ meta/runtime.yml
@@ -3,12 +3,12 @@ requires_ansible: '>=2.18.0'
 plugin_routing:
   modules:
     panos_admin:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Use 'paloaltonetworks.panos.panos_administrator' instead.
       redirect: paloaltonetworks.panos.panos_administrator
     panos_commit:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: >
           Use 'paloaltonetworks.panos.panos_commit_firewall',
@@ -16,58 +16,53 @@ plugin_routing:
           or 'paloaltonetworks.panos.panos_commit_push' instead.
       redirect: paloaltonetworks.panos.panos_commit_firewall
     panos_dag:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Use 'paloaltonetworks.panos.panos_address_group' instead.
       redirect: paloaltonetworks.panos.panos_address_group
     panos_dag_tags:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Using new modern API calls in the paloaltonetworks.panos.panos_registered_ip
       redirect: paloaltonetworks.panos.panos_registered_ip
     panos_object:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Updated to idempotent modules
       redirect: paloaltonetworks.panos.panos_address_object
     panos_object_facts:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Updating module design to network resource modules.
-    panos_nat_rule:
-      deprecation:
-        removal_version: "4.0.0"
-        warning_text: Updating module design to network resource modules.
-      redirect: paloaltonetworks.panos.panos_nat_rule2
     panos_nat_rule_facts:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Updating module design to network resource modules.
     panos_query_rules:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Querying rules is handled better by 'paloaltonetworks.panos.panos_match_rule'.
       redirect: paloaltonetworks.panos.panos_match_rule
     panos_sag:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Use 'paloaltonetworks.panos.panos_address_group' instead.
       redirect: paloaltonetworks.panos.panos_address_group
 
     panos_security_rule_facts:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Use 'paloaltonetworks.panos.panos_security_rule' with 'state=gathered'.
       redirect: paloaltonetworks.panos.panos_security_rule
 
     panos_virtual_router_facts:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Use 'paloaltonetworks.panos.panos_virtual_router' with 'state=gathered'.
       redirect: paloaltonetworks.panos.panos_virtual_router
 
     panos_zone_facts:
-      deprecation:
+      tombstone:
         removal_version: "4.0.0"
         warning_text: Use 'paloaltonetworks.panos.panos_zone' with 'state=gathered'.
       redirect: paloaltonetworks.panos.panos_zone
