--- plugins/modules/panos_virtual_router_facts.py
@@ -1,170 +1,0 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-
-#  Copyright 2019 Palo Alto Networks, Inc
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-from __future__ import absolute_import, division, print_function
-
-__metaclass__ = type
-
-DOCUMENTATION = """
----
-module: panos_virtual_router_facts
-short_description: Retrieves virtual router information
-description:
-    - Retrieves information on virtual routers from a firewall or Panorama.
-author: "Garfield Lee Freeman (@shinmog)"
-version_added: '1.0.0'
-deprecated:
-    alternative: Use M(paloaltonetworks.panos.panos_virtual_router) with I(state=gathered)
-    removed_in: '4.0.0'
-    why: Updating module design to network resource modules.
-requirements:
-    - pan-python
-    - pandevice
-notes:
-    - Panorama is supported.
-    - Check mode is not supported.
-extends_documentation_fragment:
-    - paloaltonetworks.panos.fragments.transitional_provider
-    - paloaltonetworks.panos.fragments.full_template_support
-options:
-    name:
-        description:
-            - Name of the virtual router.
-        type: str
-"""
-
-EXAMPLES = """
-# Get information on a specific virtual router
-- name: Get vr3 info
-  paloaltonetworks.panos.panos_virtual_router_facts:
-    provider: '{{ provider }}'
-    name: 'vr3'
-  register: ans
-
-# Get the config of all virtual routers
-- name: Get all virtual routers
-  paloaltonetworks.panos.panos_virtual_router_facts:
-    provider: '{{ provider }}'
-  register: vrlist
-"""
-
-RETURN = """
-spec:
-    description: The spec of the specified virtual router.
-    returned: When I(name) is specified.
-    type: complex
-    contains:
-        name:
-            description: Virtual router name.
-            type: str
-        interface:
-            description: List of interfaces
-            type: list
-        ad_static:
-            description: Admin distance for this protocol.
-            type: int
-        ad_static_ipv6:
-            description: Admin distance for this protocol.
-            type: int
-        ad_ospf_int:
-            description: Admin distance for this protocol.
-            type: int
-        ad_ospf_ext:
-            description: Admin distance for this protocol.
-            type: int
-        ad_ospfv3_int:
-            description: Admin distance for this protocol.
-            type: int
-        ad_ospfv3_ext:
-            description: Admin distance for this protocol.
-            type: int
-        ad_ibgp:
-            description: Admin distance for this protocol.
-            type: int
-        ad_ebgp:
-            description: Admin distance for this protocol.
-            type: int
-        ad_rip:
-            description: Admin distance for this protocol.
-            type: int
-vrlist:
-    description: List of virtual router specs.
-    returned: When I(name) is not specified.
-    type: list
-"""
-
-from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.paloaltonetworks.panos.plugins.module_utils.panos import (
-    get_connection,
-)
-
-try:
-    from panos.errors import PanDeviceError
-    from panos.network import VirtualRouter
-except ImportError:
-    try:
-        from pandevice.errors import PanDeviceError
-        from pandevice.network import VirtualRouter
-    except ImportError:
-        pass
-
-
-def main():
-    helper = get_connection(
-        template=True,
-        template_stack=True,
-        with_classic_provider_spec=True,
-        argument_spec=dict(
-            name=dict(),
-        ),
-    )
-    module = AnsibleModule(
-        argument_spec=helper.argument_spec,
-        supports_check_mode=True,
-        required_one_of=helper.required_one_of,
-    )
-
-    module.deprecate(
-        "Deprecated; use panos_virtual_router with state=gathered",
-        version="4.0.0",
-        collection_name="paloaltonetworks.panos",
-    )
-
-    # Verify imports, build pandevice object tree.
-    parent = helper.get_pandevice_parent(module)
-
-    name = module.params["name"]
-    if name is None:
-        try:
-            listing = VirtualRouter.refreshall(parent)
-        except PanDeviceError as e:
-            module.fail_json(msg="Failed refreshall: {0}".format(e))
-
-        module.exit_json(changed=False, vrlist=helper.describe(listing))
-
-    vr = VirtualRouter(name)
-    parent.add(vr)
-    try:
-        vr.refresh()
-    except PanDeviceError as e:
-        module.fail_json(msg="Failed refresh: {0}".format(e))
-
-    module.exit_json(changed=False, spec=helper.describe(vr))
-
-
-if __name__ == "__main__":
-    main()
